requirements.txt
streamlit
torch
torchvision
tensorflow==2.12.0
keras==2.12.0
xgboost
joblib
pillow
pandas
numpy==1.23.5
scikit-learn
